refactor: optimization
