+++ routers/card/fields_handlers/article.py
@@ -2,10 +2,10 @@ from aiogram import Router, types, F
 from aiogram.fsm.context import FSMContext
 from aiogram.utils import markdown
 
-from config import settings
+from config_data.config import settings
 from validators.card import validate_article
-from routers.card.card_handler import handle_card
-from routers.card.states import CardStates, set_input_data, Card
+from routers.card.base_handler import handle_card
+from states.states import CardStates, set_input_data, Card
 
 router = Router(name=__name__)
