+++ routers/card/fields_handlers/gn.py
@@ -3,8 +3,8 @@ from aiogram.fsm.context import FSMContext
 from aiogram.utils import markdown
 
 from validators.card import validate_gn
-from routers.card.card_handler import handle_card
-from routers.card.states import CardStates, Card, set_input_data
+from routers.card.base_handler import handle_card
+from states.states import CardStates, Card, set_input_data
 
 router = Router(name=__name__)
