+++ routers/card/fields_handlers/model.py
@@ -3,8 +3,8 @@ from aiogram.fsm.context import FSMContext
 from aiogram.utils import markdown
 
 from validators.card import validate_model
-from routers.card.card_handler import handle_card
-from routers.card.states import CardStates, set_input_data, Card
+from routers.card.base_handler import handle_card
+from states.states import CardStates, set_input_data, Card
 
 router = Router(name=__name__)
